Update example code in `README.md`
It was my mistake to convert the example code to the new style without checking it in a real app, turns out there were a few errors in the converted version.

Calling `alert` on `JSObject.global` still requires force unwrapping. Also, dynamic properties can't be updated on `JSValue` constants, so need to convert `divElement` to a `var` binding.