Supply JSKit runtime in SwiftPM resources (#193)
Due to the fact that https://github.com/apple/swift-package-manager/pull/4306 will only be available starting with Swift 5.7, we can't generate these resources with a SwiftPM plugin. I propose storing the generated code in the repository directly for now. When 5.7 is released, we can switch to a SwiftPM plugin approach.

Can be tested end-to-end with https://github.com/swiftwasm/carton/pull/335.

Co-authored-by: Yuta Saito <kateinoigakukun@gmail.com>