Start migrating imported functions to the new definition style
Use of different function names for C-name and Wasm's import name is a
bit confusing. Also use of unprefixed function names in the C code is
not a good practice. This commit starts migrating imported functions to
the new naming convention and removes duplicated function definitions
just for IDE support by using macro.