Initial WebWorkerTaskExecutor
WebWorkerTaskExecutor is an implementation of `TaskExecutor` protocol,
which is introduced by [SE-0417] since Swift 6.0. This task executor
runs tasks on Worker threads, which is useful for offloading
computationally expensive tasks from the main thread.

The `WebWorkerTaskExecutor` is designed to work with [Web Workers API]
and Node.js's [`worker_threads` module].

[SE-0417]: https://github.com/swiftlang/swift-evolution/blob/main/proposals/0417-task-executor-preference.md
[Web Workers API]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API
[`worker_threads` module]: https://nodejs.org/api/worker_threads.html