Concurrency: Remove `@Sendable` requirement from scheduling primitives
They are accessed from a single thread, so there is no need to enforce
`@Sendable` requirement on them. And also the following code is not
working with `@Sendable` requirement because the captured `JSPromise`
is not `Sendable`.

```
let promise = JSPromise(resolver: { resolver -> Void in
    resolver(.success(.undefined))
})
let setTimeout = JSObject.global.setTimeout.function!
let eventLoop = JavaScriptEventLoop(
    queueTask: { job in
        // TODO(katei): Should prefer `queueMicrotask` if available?
        // We should measure if there is performance advantage.
        promise.then { _ in
            job()
            return JSValue.undefined
        }
    },
    setTimeout: { delay, job in
        setTimeout(JSOneshotClosure { _ in
            job()
            return JSValue.undefined
        }, delay)
    }
)
```