Export `UnsafeEventLoopYield` error type
Some apps wrap instance.exports and monitor
exceptions thrown during execution of Wasm program
but `UnsafeEventLoopYield` is not something they
want to report, so they need to be able to filter them
out. However, they cannot use `UnsafeEventLoopYield` type
name because some bundlers can rename it.
We should export it to allow them not to depend on the
constructor name