BridgeJS: Improve Xcode editing experience by removing symlinks
Xcode does not handle symlinks well, especially when showing errors
in the editor. This commit removes the symlinks in the BridgeJS
plugin package and use SwiftPM's native target structure.
We still use symlinks for the actual BridgeJSTool executable target
in the main package as a build-time optimization, but it should not
affect the Xcode editing experience on the main package.