[PackageToJS] Fix flaky `timestampBasedRebuild` test by abstracting file system operations
The mtime-based change detection is not very reliable in tests due to
the coarse timestamp resolution of some file systems, especially on
artificial file operations in tests that happen in quick succession.
This change introduces an abstraction over file system operations in
MiniMake, allowing tests to use an in-memory file system with precise
control over modification times.

https://apenwarr.ca/log/20181113